Fix bug where 32-bit constants with a set high bit weren't being compiled properly on 64-bit platforms.
bk: 48ad5d6d-mjKonsQ8gsM3SmzWEVOog

+++ generic/Last.h
@@ -139,7 +139,7 @@ struct L_expr {
 	union {
 		char *string;
 		double flote;
-		int integer;
+		long integer;
 	} u;
 };
